Split up elements header file
